requirements.txt

Django==5.1.5
djangorestframework==3.15.2
gunicorn==23.0.0
requests==2.32.3
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85

